feat(sidebar): deprecate sidebarlayout component
